snap text to position, prevent some invalid builds

+++ src/gui/gui_element_types.cpp
@@ -609,14 +609,14 @@ void simple_text_element_base::render(sys::state& state, int32_t x, int32_t y) n
 			// ycentered = std::max(ycentered + state.font_collection.fonts[font_id - 1].top_adjustment(font_size),
 			// float(base_data.data.text.border_size.y));
 			ogl::render_text(state, stored_text.c_str(), uint32_t(stored_text.length()), ogl::color_modification::none,
-					float(x + text_offset), float(y + base_data.data.text.border_size.y),
+					float(x + int32_t(text_offset)), float(y + base_data.data.text.border_size.y),
 					black_text ? ogl::color3f{0.0f, 0.0f, 0.0f} : ogl::color3f{1.0f, 1.0f, 1.0f}, base_data.data.button.font_handle);
 		} else {
 			auto linesz = state.font_collection.line_height(state, base_data.data.button.font_handle);
 			auto ycentered = (base_data.size.y - linesz) / 2;
 
 			ogl::render_text(state, stored_text.c_str(), uint32_t(stored_text.length()), ogl::color_modification::none,
-					float(x + text_offset), float(y + ycentered),
+					float(x + int32_t(text_offset)), float(y + ycentered),
 					black_text ? ogl::color3f{0.0f, 0.0f, 0.0f} : ogl::color3f{1.0f, 1.0f, 1.0f}, base_data.data.text.font_handle);
 		}
 	}
